+++ doc/source/user_guide/groupby.rst
@@ -320,15 +320,6 @@ number:
 
    s.groupby(level="second").sum()
 
-The aggregation functions such as ``sum`` will take the level parameter
-directly. Additionally, the resulting index will be named according to the
-chosen level:
-
-.. ipython:: python
-   :okwarning:
-
-   s.sum(level="second")
-
 Grouping with multiple levels is supported.
 
 .. ipython:: python
